move CameraPropertiesImpl to its own file; make CameraPropertiesImpl public
